Restructured files in Components, added comments
