+++ blueview/src/Components/DepartmentPage.js
@@ -1,10 +1,10 @@
 import React from 'react';
-import Post from './Post.js';
 import baseUrl from '../Utils/config'
 import axios from 'axios';
+import PostFeed from 'PostComponents.js';
 import './DepartmentPage.css';
-import './NewPost.css';
 
+// Topmost header ofor the page
 class Header extends React.Component {
   constructor(props) {
     super(props);
@@ -17,6 +17,7 @@ class Header extends React.Component {
     this.signIn = this.signIn.bind(this);
     this.signOut = this.signOut.bind(this);
   }
+
   search(event) {
     alert("Search");
   }
@@ -29,6 +30,7 @@ class Header extends React.Component {
     alert("Sign Out");
   }
 
+  // Consists of the logo, a search placeholder, and the sign in/out button (not yet linked)
   render() {
     const loggedIn = this.state.loggedIn;
     let accountControl;
@@ -48,6 +50,7 @@ class Header extends React.Component {
   }
 }
 
+// Displays the aggregated ratings for the department
 class DepartmentRatings extends React.Component {
   constructor(props) {
     super(props);
@@ -71,6 +74,7 @@ class DepartmentRatings extends React.Component {
   }
 }
 
+// Main header for any department page.
 class DepartmentHeader extends React.Component {
   constructor(props) {
     super(props);
@@ -78,7 +82,7 @@ class DepartmentHeader extends React.Component {
     
   }
   
-  
+  // Make API call to get department information after mounting. Updates state which triggers a rerender.
   componentDidMount() {
     axios.get(`${baseUrl}/department/${this.props.id}`)
       .then(res => {
@@ -91,6 +95,7 @@ class DepartmentHeader extends React.Component {
       })
   }
 
+  // Consists of department name and address with a DepartmentRatings component
   render() {
     return (
       <div className="departmentHeaderMain">
@@ -107,177 +112,13 @@ class DepartmentHeader extends React.Component {
   }
 }
 
-class PostControls extends React.Component {
-  constructor(props) {
-    super(props);
-  }
-
-  render() {
-    return (
-      <div className="postControlsContainer">
-        <button className="createNewPostButton" onClick={this.props.showNewPost}>
-          New Post
-        </button>
-      </div>
-    );
-  }
-}
-
-class NewPost extends React.Component {
-  constructor(props) {
-    super(props);
-
-    this.state = {
-      title: '',
-      date: '',
-      text: '',
-      attitude: '',
-      communication: '',
-      efficiency: '',
-      fairness: '',
-      safety: ''
-    }
-
-    this.handleChange = this.handleChange.bind(this);
-    this.handleSubmit = this.handleSubmit.bind(this);
-  }
-
-  handleChange(event) {
-    this.setState({
-      [event.target.name]: event.target.value
-    });
-  }
-
-  handleSubmit(event) {
-    event.preventDefault();
-    alert("Title: " + this.state.title +
-          "\nDate: " + this.state.date +
-          "\nText: " + this.state.text);
-
-    this.props.cancelPost(); // Hide new post form
-  }
-
-  render() {
-    return (
-      <div className="newPostContainer">
-        <form onSubmit={this.handleSubmit}>
-          <div className="newPostRow">
-            <div className="newPostItem">
-              <label className="newPostLabel">Title:</label>
-              <input className="newPostTitle" type="text" value={this.state.title} name="title" onChange={this.handleChange} />
-            </div>
-
-            <div className="newPostItem">
-              <label className="newPostLabel">Date of event:</label>
-              <input className="newPostDate" type="date" name="date" onChange={this.handleChange} />
-            </div>
-
-            <div className="newPostItem">
-              <label className="newPostLabel">Tag:</label>
-              <select className="newPostTag" type="select" name="tag" onChange={this.handleChange}>
-                <option value="Traffic Stop">Traffic Stop</option>
-                <option value="Emergency Response">Emergency Response</option>
-                <option value="Noise Complaint">Noise Complaint</option>
-                <option value="Domestic Dispute">Domestic Call</option>
-              </select>
-            </div>
-          </div>
-
-          
-          <label className="newPostLabel">Post text:</label>
-          <textarea className="newPostText" value={this.state.text} name="text" onChange={this.handleChange} />
-
-          <div className="ratingRow">
-            <div className="newPostItem">
-              <label className="ratingLabel">Attitude:</label>
-              <input className="ratingEntry" type="number" min="1" max="5" value={this.state.attitude} name="attitude" onChange={this.handleChange} />
-            </div>
-            <div className="newPostItem">
-              <label className="ratingLabel">Communication:</label>
-              <input className="ratingEntry" type="number" min="1" max="5" value={this.state.communication} name="communication" onChange={this.handleChange} />
-            </div>
-            <div className="newPostItem">
-              <label className="ratingLabel">Efficiency:</label>
-              <input className="ratingEntry" type="number" min="1" max="5" value={this.state.efficiency} name="efficiency" onChange={this.handleChange} />
-            </div>
-          </div>
-
-          <div className="ratingRow">
-            <div className="newPostItem">
-              <label className="ratingLabel">Fairness:</label>
-              <input className="ratingEntry" type="number" min="1" max="5" value={this.state.fairness} name="fairness" onChange={this.handleChange} />
-            </div>
-            <div className="newPostItem">
-              <label className="ratingLabel">Safety:</label>
-              <input className="ratingEntry" type="number" min="1" max="5" value={this.state.safety} name="safety" onChange={this.handleChange} />
-            </div>
-          </div>
-
-          <div className="newPostControls">
-            <button className="cancelPost" onClick={this.props.cancelPost}>Cancel</button>
-            <input className="submitPost" type="submit" value="Post" />
-          </div>
-        </form>
-      </div>
-    );
-  }
-}
-
-class PostFeed extends React.Component {
-  constructor(props) {
-    super(props);
-
-    this.state = {
-      creatingPost: false,
-      posts: []
-    }
-
-    this.showNewPost = this.showNewPost.bind(this);
-    this.cancelPost = this.cancelPost.bind(this);
-  }
-
-  componentDidMount() {
-    axios.get(`${baseUrl}/department/${this.props.id}/post/all`)
-      .then(res => {
-        const data = res.data.department_posts;
-        console.log("PostFeed Mount");
-        console.log(data);
-        this.setState({posts: data});
-      });
-  }
-
-  showNewPost() {
-    this.setState({
-      creatingPost: true
-    });
-  }
-
-  cancelPost() {
-    this.setState({
-      creatingPost: false
-    });
-  }
-
-  render() {
-    const { posts } = this.state;
-    return (
-      <div className="postFeedContainer">
-        <PostControls showNewPost={this.showNewPost} />
-        {this.state.creatingPost && (<NewPost cancelPost={this.cancelPost} />)}
-        {posts.map(post => (
-          <Post title={post.title} date={post.date} 
-          user={post.user} text={post.text}  />
-        ))}
-      </div>
-    );
-  }
-}
-
+// Everything below the header in a department page.
 class DepartmentContent extends React.Component {
   constructor(props) {
     super(props);
   }
 
+  // Currently only contains a PostFeed component
   render() {
     return (
       <div className="departmentContentMain">
@@ -289,10 +130,13 @@ class DepartmentContent extends React.Component {
   }
 }
 
+// The entire department page.
 class DepartmentPage extends React.Component {
   constructor(props) {
     super(props);
   }
+
+  // Consists of the topmost header, the department-specific header, and the department content
   render() {
     return (
       <div className="departmentPageMain">
